Resolve conflicts in tests/integration/test_ssl_cert_authentication
caused by master's PR #108472 (SSL certificate SAN wildcard matching),
which regenerated the whole certificate set and renumbered clients up to
`client12` plus `client_far_future`, colliding with this branch's
`client7`.

Resolution:
- Take master's consistent certificate set as-is (all signed by master's
  CA) so the existing wildcard / session_log tests keep passing.
- Move this branch's EMAIL SAN certificate to a fresh `client13`
  (`EMAIL:alice@example.com`), signed by master's CA, and update
  `generate_certs.sh` accordingly.
- Keep master's `users_with_ssl_auth.xml` users and add back the `alice`
  user that authenticates by `EMAIL:alice@example.com`.
- Take master's `test.py` and re-append `test_x509_san_email_support`,
  switched from `client7` to `client13`.
- `src/Common/Crypto/X509Certificate.cpp` auto-merged: `extractAllSubjects`
  keeps the new `GEN_EMAIL` handling on top of master's changes.

+++ src/BridgeHelper/XDBCBridgeHelper.cpp
@@ -15,7 +15,7 @@ Poco::URI XDBCBridgeHelper<T>::createBaseURI() const
 {
     Poco::URI uri;
     uri.setHost(bridge_host);
-    uri.setPort(bridge_port);
+    uri.setPort(static_cast<uint16_t>(bridge_port));
     uri.setScheme("http");
     uri.addQueryParameter("use_connection_pooling", toString(use_connection_pooling));
     return uri;
